requirements.txt
werkzeug
flask
Pillow==6.2.0
